some rearranging and added restricted pickling
